Clock app
Clocker app is daily task management / planner app. The user will be
able to add new tasks, and each task / event will be able to produce
alerts and notifications throughout the day. Each task can be given
specific information such as time and location ofeach task, and
clicking on a task that contains a phone number, an email, or a map
location will
open the phone, email or map apps on the phone with the given
information. We don’t have to remember a lot of things and can use this
app for alarm notifications and i have used Alarm system service
provided by Android to set alarms and Notification system service for
notifications.